8261075: Create stubRoutines.inline.hpp with SafeFetch implementation
Backport-of: b955f85e03bafe8ce39677d0af06bf1ceb7e2cbb
